+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -6,13 +6,13 @@ import java.time.LocalDate;
 
 @Data
 public class Film {
-    private int id;
+    private long id;
     private String name;
     private String description;
     private LocalDate releaseDate;
     private int duration;
 
-    public Film(int id, String name, String description, LocalDate releaseDate, int duration) {
+    public Film(long id, String name, String description, LocalDate releaseDate, int duration) {
         this.id = id;
         this.name = name;
         this.description = description;
