+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -6,13 +6,13 @@ import java.time.LocalDate;
 
 @Data
 public class User {
-    private int id;
+    private long id;
     private String email;
     private String login;
     private String name;
     private LocalDate birthday;
 
-    public User(int id, String email, String login, String name, LocalDate birthday) {
+    public User(long id, String email, String login, String name, LocalDate birthday) {
         this.id = id;
         this.email = email;
         this.login = login;
